🎨 Base class
